requirements.txt
torch==2.1.0+cu118
torchvision==0.16.0+cu118

absl-py==2.0.0
accelerate==0.27.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==5.0.1
attrs==23.2.0
aura-sr==0.0.4
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
blessed==1.21.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cholespy==1.0.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
coloredlogs==15.0.1
comm==0.2.1
compel==2.0.3
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.14
diffusers==0.29.0
docker-pycreds==0.4.0
easydict==1.10
einops==0.8.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastjsonschema==2.19.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.7.0
fsspec==2023.9.2
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.59.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.4
igl @ file:///home/conda/feedstock_root/build_artifacts/igl_1611531307820/work
imageio==2.31.5
imageio-ffmpeg==0.4.9
imath==0.0.2
img2texture==1.1.2.post2
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.2
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kaolin @ git+https://github.com/NVIDIAGameWorks/kaolin@5406915c3a020596384b1a36679ff68afd04d455
kiwisolver==1.4.5
kornia==0.7.0
largesteps @ file:///home/hengyumeng/large-steps-pytorch
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.15.2
llvmlite==0.42.0
loguru==0.7.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mvdream @ git+https://github.com/bytedance/MVDream@618f84476c88e04adaf588133be2830142b9956d
mypy-extensions==1.0.0
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
ninja==1.11.1.1
noise==1.2.2
numba==0.59.1
numpy==1.24.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@78528e683210dfaa1be57e3c65aa37d3b36c6644
nvidia-ml-py==13.580.65
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.18.0
open3d==0.18.0
open_clip_torch==2.31.0
opencv-python-headless==4.9.0.80
OpenEXR==3.2.4
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.1
platformdirs==4.2.0
plotly==5.22.0
pooch==1.8.1
portalocker==2.10.1
prompt-toolkit==3.0.43
propcache==0.3.2
protobuf==4.24.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pydantic==2.10.2
pydantic_core==2.27.1
PyGLM==2.7.0
pygltflib==1.16.2
Pygments==2.17.2
PyJWT==2.8.0
PyMatting==1.1.12
PyMCubes==0.1.6
pymeshlab==2022.2.post4
pyparsing==3.1.1
pyproject==1.3.1
pyquaternion==0.9.9
pyrallis==0.3.1
python-dateutil==2.8.2
pytorch-lightning==2.5.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@89653419d0973396f3eff1a381ba09a07fffc2ed
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.35.1
regex==2023.10.3
rembg==2.0.56
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
retrying==1.3.4
rich==14.1.0
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.5.0
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073865807/work
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tabulate==0.9.0
taming-transformers-rom1504==0.0.6
tenacity==8.3.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.4
termcolor==2.3.0
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==1.0.15
tokenizers==0.15.2
torch==2.1.0+cu118
torch-ema==0.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torchmetrics==1.8.1
torchvision==0.16.0+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
trimesh==3.23.5
triton==2.1.0
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
urllib3==2.0.6
usd-core==23.5
wandb==0.18.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xatlas==0.0.7
yacs==0.1.8
yarl==1.20.1
zhipuai==2.1.5.20230904
zipp==3.17.0
